Feature/v2.0

+++ app/Http/Controllers/RepositoryController.php
@@ -4,17 +4,35 @@ namespace App\Http\Controllers;
 
 use App\Models\Repository;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Auth;
 
 class RepositoryController extends Controller
 {
     public function index()
     {
-        $repositories = Repository::with('project')->get();
+        if (!Auth::user()->is_admin) {
+            $repositories = Repository::select('repositories.id', 'repositories.name as repository_name', 'projects.name as project_name')
+            ->leftJoin('user_repositories', 'user_repositories.repository_id', '=', 'repositories.id')
+            ->leftJoin('projects', 'projects.id', '=', 'repositories.project_id')
+            ->where('user_repositories.user_id', Auth::user()->id)
+            ->get();
+            return response()->json($repositories);
+        }
+        $repositories = Repository::select('repositories.id', 'repositories.name as repository_name', 'projects.name as project_name')
+        ->leftJoin('projects', 'projects.id', '=', 'repositories.project_id')
+        ->get();
+
         return response()->json($repositories);
     }
 
     public function store(Request $request)
     {
+        if (!Auth::user()->is_admin) {
+            return response()->json([
+                "success" => false,
+                "message" => "Access denied"
+            ], 403);
+        }
         $validated = $request->validate([
             'project_id' => 'required|exists:projects,id',
             'name' => 'required|string|max:255',
@@ -34,6 +52,13 @@ class RepositoryController extends Controller
 
     public function update(Request $request, $id)
     {
+        if (!Auth::user()->is_admin) {
+            return response()->json([
+                "success" => false,
+                "message" => "Access denied"
+            ], 403);
+        }
+
         $validated = $request->validate([
             'project_id' => 'sometimes|required|exists:projects,id',
             'name' => 'sometimes|required|string|max:255',
@@ -48,6 +73,12 @@ class RepositoryController extends Controller
 
     public function destroy($id)
     {
+        if (!Auth::user()->is_admin) {
+            return response()->json([
+                "success" => false,
+                "message" => "Access denied"
+            ], 403);
+        }
         $repository = Repository::findOrFail($id);
         $repository->delete();
